requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
Flask==3.1.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tensorflow==2.18.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
google-generativeai==0.8.3
google-cloud-secret-manager==2.21.1